Move test utils file as it errors in CI
